Improved "allowSign" method on the NoInteractions Module(ServerPlayerEntity->ClientPlayerEntity)
