+++ src/main/java/io/github/itzispyder/clickcrystals/mixins/MixinClientPlayerEntity.java
@@ -4,9 +4,11 @@ package io.github.itzispyder.clickcrystals.mixins;
 import io.github.itzispyder.clickcrystals.Global;
 import io.github.itzispyder.clickcrystals.modules.Module;
 import io.github.itzispyder.clickcrystals.modules.modules.anchoring.ElytraSwitch;
+import io.github.itzispyder.clickcrystals.modules.modules.misc.NoInteractions;
 import io.github.itzispyder.clickcrystals.util.minecraft.HotbarUtils;
 import io.github.itzispyder.clickcrystals.util.minecraft.InteractionUtils;
 import io.github.itzispyder.clickcrystals.util.minecraft.PlayerUtils;
+import net.minecraft.block.entity.SignBlockEntity;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.entity.effect.StatusEffects;
 import org.spongepowered.asm.mixin.Mixin;
@@ -30,6 +32,13 @@ public class MixinClientPlayerEntity implements Global {
             if (HotbarUtils.isHolding(ELYTRA)) {
                 InteractionUtils.inputUse();
             }
+        }
+    }
+    @Inject(method = "openEditSignScreen", at = @At("HEAD"), cancellable = true)
+    private void cancelOpenContainer(SignBlockEntity sign, boolean front, CallbackInfo ci) {
+        NoInteractions noInteractions = Module.get(NoInteractions.class);
+        if (noInteractions.isEnabled() && !noInteractions.allowSign.getVal()) {
+            ci.cancel();
         }
     }
 }
