+++ src/main/resources/clickcrystals.mixins.json
@@ -42,7 +42,6 @@
     "MixinLivingEntity",
     "MixinMinecraftClient",
     "MixinPlayerEntity",
-    "MixinPlayerInventory",
-    "MixinServerPlayerEntity"
+    "MixinPlayerInventory"
   ]
 }
